+++ lib/pleroma/web/pleroma_api/controllers/account_controller.ex
@@ -106,7 +106,7 @@ defmodule Pleroma.Web.PleromaAPI.AccountController do
     users =
       user
       |> User.endorsed_users_relation(_restrict_deactivated = true)
-      |> Pleroma.Pagination.fetch_paginated(Map.put(params, :skip_order, true))
+      |> fetch_paginated_endorsements(params)
 
     conn
     |> add_link_headers(users)
@@ -116,6 +116,16 @@ defmodule Pleroma.Web.PleromaAPI.AccountController do
       as: :user,
       embed_relationships: embed_relationships?(params)
     )
+  end
+
+  defp fetch_paginated_endorsements(user, %{shuffle: true} = params) do
+    user
+    |> Pleroma.Pagination.fetch_paginated(Map.put(params, :shuffle, true))
+  end
+
+  defp fetch_paginated_endorsements(user, params) do
+    user
+    |> Pleroma.Pagination.fetch_paginated(Map.put(params, :skip_order, true))
   end
 
   @doc "POST /api/v1/pleroma/accounts/:id/subscribe"